% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateCodingVcf.R
\name{annotateCodingVcf}
\alias{annotateCodingVcf}
\title{Annotate a VCF object}
\usage{
annotateCodingVcf(vcf)
}
\arguments{
\item{vcf}{A \code{VCF} object belonging to the \code{VCF} class from the
\code{VariantAnnotation} package. Can be loaded by using the \code{readVcf()}
function on a suitable VCF file. The VCF file must follow the VCF format
specifications. See \code{?VariantAnnotation::readVcf()} for more details.}
}
\value{
a \code{GRanges} object with a row for each variant-transcript match. For
more information about the information stored in the columns, see the \verb{See Also} section.
}
\description{
Annotates a VCF object using the \code{predictCoding()} function from
VariantAnnotation using data from the \code{TxDb.Hsapiens.UCSC.hg19.knownGene}
package. Specifically, it adds transcript-level information about coding
sequence mutations, such as transcript ID, exon number, distance of the
mutation site from the transcription start site (excluding introns and UTRs),
etc. Mutations in non-coding regions are excluded.
}
\examples{
annotatedVcf <- annotateCodingVcf(mutationsVcf)
annotatedVcf
}
\seealso{
To obtain more information about the metadata columns in the
returned \code{GRanges} object, see \code{\link[VariantAnnotation:predictCoding-methods]{VariantAnnotation::predictCoding()}} which
is used by this function to obtain the annotated \code{GRanges} object.
}
