% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnedByDistancePlot.R
\name{calculateNormDistances}
\alias{calculateNormDistances}
\title{Calculating Normalized Distances for Mutations}
\usage{
calculateNormDistances(codingVcfDf)
}
\arguments{
\item{codingVcfDf}{}
}
\value{
A dataframe containing added columns with transcript length and normalized mutation distance.
}
\description{
Calculating Normalized Distances for Mutations
}
