% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnedByDistancePlot.R
\name{generateBinnedPlotV2}
\alias{generateBinnedPlotV2}
\title{Generating Plot of Mutation Data Binned by Distance from Start Site}
\usage{
generateBinnedPlotV2(codingVcfDf, nBins, type = "CONSEQUENCE")
}
\arguments{
\item{codingVcfDf}{A `VCF` object generated using `annotateVCF()` containing
coding variants with metadata, including a `CDSLOC.start` column indicating
the position of the mutation in the coding sequence. Must also contain a
column for the user-defined variant type.}

\item{nBins}{An integer specifying the number of bins to divide the coding
region into. Default is 10.}

\item{type}{A string specifying the property to group the variants by in each
stacked barplot.}
}
\value{
A bar plot showing the distribution of variant
counts across bins grouped by their distance from the 5' end of the coding
sequence.
}
\description{
Generating Plot of Mutation Data Binned by Distance from Start Site
}
