% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVCF.R
\name{loadVcf}
\alias{loadVcf}
\title{Load a VCF file into R}
\usage{
loadVcf(pathToVcf, genome)
}
\arguments{
\item{pathToVcf}{A string representing the file path to the VCF file. The VCF file must have mutations aligned using the hg19 reference genome; other genomes are not supported at the moment.}

\item{genome}{A string specifying the reference genome version (e.g. "hg19") (Note: right now, this is the only version that is supported. Additional functionality for supporting other genome types will be added in the future.)}
}
\value{
A `VCF` object containing the variants from the VCF file.
}
\description{
A function that loads a VCF (Variant Call Format) file, specified by the file path, into R.
It returns a VCF object, which can be further used for variant analysis.
}
\examples{
\dontrun{
vcfData <- loadVCF("/path/to/file.vcf", "hg19")
}

}
