% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutationsVcf}
\alias{mutationsVcf}
\title{Mutations detected by sequencing a human genome}
\format{
An object of class \code{CollapsedVCF} with 10376 rows and 5 columns.
}
\source{

}
\usage{
mutationsVcf
}
\description{
A VCF object found in the VariantAnnotation package, obtained by using the `vcf_filename <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")` command followed by the `readVcf(vcf_filename)` command. It contains mutations from the genome sequence of chr22 of an individual.
}
\examples{
\dontrun{
 mutationsVcf
}

}
\keyword{datasets}
