% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVcf.R
\name{loadVcf}
\alias{loadVcf}
\title{Load a VCF file into R}
\usage{
loadVcf(pathToVcf)
}
\arguments{
\item{pathToVcf}{A string representing the file path to the VCF file. The VCF file must have mutations aligned using the hg19 reference genome; other genomes are not supported at the moment.}
}
\value{
A `VCF` object containing the variants from the VCF file. Seqnames are ensured to have the 'chr' prefix.
}
\description{
A function that loads a VCF (Variant Call Format) file, specified by the file path, into R.
It returns a VCF object, which can be further used for variant analysis. It also adds the 'chr' prefix to
chromosome names if it hasn't been added already, to make downstream steps compliant with preconditions.
}
\examples{
\dontrun{
vcfData <- loadVCF("/path/to/file.vcf", "hg19")
}

}
