% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVCF.R
\name{loadVcf}
\alias{loadVcf}
\title{Load a VCF file into R}
\usage{
loadVcf(pathToVcf, genome)
}
\arguments{
\item{pathToVcf}{A string representing the file path to the VCF file.}

\item{genome}{A string specifying the reference genome version (e.g. "hg19")}
}
\value{
A `VCF` object containing the variants from the VCF file.
}
\description{
This function loads a VCF (Variant Call Format) file, specified by the file path, into R using the `VariantAnnotation` package.
It returns a VCF object, which can be further used for variant analysis.
}
\examples{
\dontrun{
vcfData <- loadVCF("/path/to/file.vcf", "hg19")
}
}
