% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutationsVcf}
\alias{mutationsVcf}
\title{Mutations detected by sequencing}
\format{
A matrix with columns:
\describe{
 \item{Condition.1}{First condition, under which least light was provided
   to plants.}
 \item{Condition.2}{Second condition, under which most light was provided to plants.}
 \item{Condition.3}{Third condition, the control.}
}
}
\source{
University of Guelph, Canada.
}
\usage{
mutationsVcf
}
\description{
An RNAseq experiment conductd using bean plants from 2016 in Canada.
}
\examples{
\dontrun{
 mutationsVcf
}
}
\keyword{datasets}
